Add files via upload
Welcome to ChemCourseMarketplace, your go-to destination for a comprehensive and interactive chemistry education experience. This open-source project is a user-friendly, feature-rich platform designed to facilitate the sale and distribution of high-quality chemistry courses. Whether you're a passionate chemistry educator looking to share your knowledge or a dedicated learner eager to explore the fascinating world of chemistry, this repository is your one-stop solution.